⠀

+++ 1-ano/TPA/java/revisao/calculo/src/main/java/com/mycompany/calculo/main.java
@@ -11,8 +11,8 @@ import javax.swing.JOptionPane;
 public class main {
     public static void main(String[] args) {
         double x = 0, n = 0, res = 0;
-        x = Integer.parseInt(JOptionPane.showInputDialog("Digite o 1º número"));
-        n = Integer.parseInt(JOptionPane.showInputDialog("Digite o 2º número"));
+        x = Double.parseDouble(JOptionPane.showInputDialog("Digite o 1º número"));
+        n = Double.parseDouble(JOptionPane.showInputDialog("Digite o 2º número"));
         res = (Math.pow((x*n), 2));
         JOptionPane.showMessageDialog(null, "O resultado é: " + res);
     }
